EDU-447 Move storage helper methods in utils folder module
